updated redirection of signin, signup, and home components, organised redux files and folders together
