+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import './styles/index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import {Store} from './store';
+import {Store} from './redux/store';
 import {Provider} from 'react-redux';
 
 ReactDOM.render(
