+++ src/redux/reducers/rootReducer.js
@@ -1,5 +1,5 @@
 import { combineReducers } from "redux";
-import {updateReducer, getData} from './';
+import {updateReducer, getData} from '.';
 
 export default combineReducers({
     updateReducer,
